Rename "tag-import" mode to "resume-import."
This is a more accurate name, plus it matches the "resume-update" mode we already have.

This commit ALSO addresses a pretty big bug wherein if there _were_ problems during the import process, we'd sort of silently ignore them and not tell the user that their code wouldn't be tagged correctly.