Fix 'non-existent git branch' error message during resume-update.
When gathering a list of branches, we actually want all the branch names, in both the local and remotes. This function was only providing the list of local branches, so if the user hadn't checked out the upstream branch (and thus set up a tracking branch), they would receive this error message.